Revert "LPS-52177 simplify user deletion"
This reverts commit e93894c5a7c9ea88bde61dab4ab164a87eff6fd2.

+++ portal-impl/src/com/liferay/portal/service/impl/CompanyLocalServiceImpl.java
@@ -1352,12 +1352,18 @@ public class CompanyLocalServiceImpl extends CompanyLocalServiceBaseImpl {
 
 				@Override
 				public void performAction(User user) throws PortalException {
-					userLocalService.deleteUser(user.getUserId());
+					if (!user.isDefaultUser()) {
+						userLocalService.deleteUser(user.getUserId());
+					}
 				}
 
 			});
 
 		userActionableDynamicQuery.performActions();
+
+		User defaultUser = userLocalService.getDefaultUser(companyId);
+
+		userLocalService.deleteUser(defaultUser);
 
 		// Groups
